change variable types from int to long int

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,18 +8,18 @@
 
 int main(void)
 {
-	int a = 1, b = 2, c, i;
+	long int a = 1, b = 2, c, i;
 
-	printf("%d, %d, ", a, b);
+	printf("%ld, %ld, ", a, b);
 
 	for (i = 0; i < 48; i++)
 	{
 		c = a + b;
 
 		if (i == 47)
-			printf("%d", c);
+			printf("%ld", c);
 		else
-			printf("%d, ", c);
+			printf("%ld, ", c);
 
 		a = b;
 		b = c;
